% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectMM10.R
\docType{class}
\name{TrenaProjectMM10-class}
\alias{TrenaProjectMM10-class}
\alias{.TrenaProjectMM10}
\alias{TrenaProjectMM10}
\title{TrenaProjectMM10-class}
\usage{
TrenaProjectMM10(quiet = TRUE)
}
\value{
An object of the TrenaProjectMM10 class
}
\description{
Expression, variant and covariate data for the genes of interest (perhaps unbounded) for pre-term birth studies
}
